More renaming & consistency
