+++ examples/deep_research_multi_agent/agents/top_urls_selector_agent.py
@@ -1,9 +1,9 @@
 from typing import List
 import instructor
 import openai
-from pydantic import BaseModel, Field, HttpUrl
+from pydantic import Field, HttpUrl
 from atomic_agents.agents.base_agent import BaseIOSchema, BaseAgent, BaseAgentConfig
-from atomic_agents.lib.components.system_prompt_generator import SystemPromptGenerator, SystemPromptInfo
+from atomic_agents.lib.components.system_prompt_generator import SystemPromptGenerator
 
 class TopUrlsSelectorInputSchema(BaseIOSchema):
     user_input: str = Field(..., description='The user input or question.')
@@ -18,20 +18,18 @@ top_urls_selector_agent = BaseAgent(
         client=instructor.from_openai(openai.OpenAI()), 
         model='gpt-3.5-turbo',
         system_prompt_generator=SystemPromptGenerator(
-            SystemPromptInfo(
-                background=[
-                    "You are an intelligent URL selection expert.",
-                    "Your task is to select the best URLs from a list of search results that are most relevant to the user input or question."
-                ],
-                steps=[
-                    "You will receive the user input or question, the list of search results, and the number of best URLs to select.",
-                    "Analyze the search results and select the best URLs that are most relevant to the user input or question.",
-                ],
-                output_instructions=[
-                    "Ensure each selected URL is highly relevant to the user input or question.",
-                    "Ensure the selected URLs are diverse yet as relevant as possible."
-                ]
-            )
+            background=[
+                "You are an intelligent URL selection expert.",
+                "Your task is to select the best URLs from a list of search results that are most relevant to the user input or question."
+            ],
+            steps=[
+                "You will receive the user input or question, the list of search results, and the number of best URLs to select.",
+                "Analyze the search results and select the best URLs that are most relevant to the user input or question.",
+            ],
+            output_instructions=[
+                "Ensure each selected URL is highly relevant to the user input or question.",
+                "Ensure the selected URLs are diverse yet as relevant as possible."
+            ]
         ),
         input_schema=TopUrlsSelectorInputSchema,
         output_schema=TopUrlsSelectorOutputSchema
